#2 Fix: entity로 리턴하는 문제 해결, 유저 아이디 long으로 변경

+++ cookiee-server/src/main/java/com/cookiee/cookieeserver/controller/EventController.java
@@ -52,7 +52,7 @@ public class EventController {
     @ResponseBody
     @PostMapping(value = "/event/{userId}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public BaseResponseDto<Event> saveEvent(
-            @PathVariable int userId, HttpServletRequest request,
+            @PathVariable Long userId, HttpServletRequest request,
             @RequestParam(value = "images") List<MultipartFile> imageUrl,
             EventRegisterRequestDto eventRegisterRequestDto) throws IOException {
         Event event;
