+++ cookiee-server/src/main/java/com/cookiee/cookieeserver/domain/User.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class User extends BaseTimeEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)  // auto_increment
-    private int userId;  // 사용자 아이디
+    private Long userId;  // 사용자 아이디
 
     @Column(nullable = false, length = 10)
     private String nickname;  // 사용자가 설정한 닉네임
